[Chore] solved deploy problem
