+++ .eslintrc.json
@@ -1,7 +1,6 @@
 {
   "extends": [
     "next/core-web-vitals",
-    "prettier",
     "eslint:recommended",
     "plugin:prettier/recommended"
   ],
