+++ package.json
@@ -15,11 +15,11 @@
     "@types/cookie": "^0.6.0",
     "@types/lodash.debounce": "^4.0.9",
     "axios": "^1.6.8",
+    "chart.js": "^4.4.3",
     "cookie": "^0.6.0",
+    "date-fns": "^3.6.0",
     "json-server": "1.0.0-beta.0",
     "lodash.debounce": "^4.0.8",
-    "chart.js": "^4.4.3",
-    "date-fns": "^3.6.0",
     "next": "14.2.3",
     "next-auth": "^4.24.7",
     "react": "^18",
@@ -38,6 +38,7 @@
     "@types/node": "^20.12.11",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@typescript-eslint/eslint-plugin": "^7.11.0",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
     "eslint-config-prettier": "^9.1.0",
